added authentication and authorization functionality
created /auth.js
created /passport.js

+++ server.js
@@ -11,6 +11,12 @@ const express = require('express'),
 mongoose.connect('mongodb://127.0.0.1:27017/myflix', {useNewUrlParser: true, useUnifiedTopology: true });
 
 app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+
+let auth = require('./auth')(app);
+
+const passport = require('passport');
+require('./passport');
 
 
 
@@ -41,7 +47,7 @@ app.post('/users', (req, res) => {
     });
 });
 
-
+/*
 // Get all users
 app.get('/users', (req, res) => {
     Users.find()
@@ -53,10 +59,15 @@ app.get('/users', (req, res) => {
             res.status(500).send('Error: ' + err);
         });
 });
+*/
 
 
 // Get Specific User
-app.get('/users/:UserName', (req, res) => {
+app.get('/users/:UserName', passport.authenticate('jwt', { session: false } ), (req, res) => {
+    if (req.user.UserName !== req.params.UserName) {
+        return res.status(400).send('Permission denied');
+    }
+
     Users.findOne({ UserName: req.params.UserName })
         .then((user) => {
             if (user) {
@@ -75,7 +86,11 @@ app.get('/users/:UserName', (req, res) => {
 
 //Update User Name
 
-app.put('/users/:UserName', (req, res) => {
+app.put('/users/:UserName', passport.authenticate('jwt', { session: false }), (req, res) => {
+    if (req.user.UserName !== req.params.UserName) {
+        return res.status(400).send('Permission denied');
+    }
+
     Users.findOneAndUpdate({ UserName: req.params.UserName }, {
         $set: 
             {
@@ -97,7 +112,11 @@ app.put('/users/:UserName', (req, res) => {
 
 
 //Add to fav movies
-app.post('/users/:UserName/movies/:MovieID', (req, res) => {
+app.post('/users/:UserName/movies/:MovieID', passport.authenticate('jwt', { session: false }), (req, res) => {
+    if (req.user.UserName !== req.params.UserName) {
+        return res.status(400).send('Permission denied')
+    }
+
     Users.findOneAndUpdate({ UserName: req.params.UserName }, 
         { $push: { FavoriteMovies: req.params.MovieID }
     },
@@ -113,7 +132,11 @@ app.post('/users/:UserName/movies/:MovieID', (req, res) => {
 
 
 //DELETE from fav movies
-app.delete('/users/:UserName/movies/:MovieID', (req, res) => {
+app.delete('/users/:UserName/movies/:MovieID', passport.authenticate('jwt', { session: false }), (req, res) => {
+    if (req.user.UserName !== req.params.UserName) {
+        return res.status(400).send('Permission denied')
+    }
+
     Users.findOneAndUpdate( {UserName: req.params.UserName} ,{
         $pull: { FavoriteMovies: req.params.MovieID }
     },
@@ -129,7 +152,11 @@ app.delete('/users/:UserName/movies/:MovieID', (req, res) => {
 
 
 //DELETE user
-app.delete('/users/:UserName', (req, res) => {
+app.delete('/users/:UserName', passport.authenticate('jwt', { session: false} ), (req, res) => {
+    if (req.user.UserName !== req.params.UserName) {
+        return res.status(400).send('Permission denied');
+    }
+
     Users.findOneAndRemove( { UserName: req.params.UserName })
         .then((user) => {
             if (!user) {
@@ -146,7 +173,7 @@ app.delete('/users/:UserName', (req, res) => {
 
 
 //READ list of movies
-app.get('/movies', (req, res) => {
+app.get('/movies', passport.authenticate('jwt', { session: false }), (req, res) => {
     Movies.find()
     .then( (movies) => {
         res.status(200).json(movies)
@@ -158,7 +185,7 @@ app.get('/movies', (req, res) => {
 })
 
 //READ specific movie
-app.get('/movies/:Title', (req, res) => {
+app.get('/movies/:Title', passport.authenticate('jwt', { session: false }), (req, res) => {
     Movies.findOne( { Title: req.params.Title } )
         .then((movie) => {
             if (movie) {
@@ -171,10 +198,10 @@ app.get('/movies/:Title', (req, res) => {
             console.error(err);
             res.status(500).send('Err: ' + err);
         })
-})
+}) 
 
 // READ genre description
-app.get('/movies/genre/:genreName', (req, res) => {
+app.get('/movies/genre/:genreName', passport.authenticate('jwt', { session: false }), (req, res) => {
     Movies.findOne( { "Genre.Name": req.params.genreName } )
         .then((genre) => {
             if (genre) {
